refactor(monorepo): extract @sudowork/contracts shared package
Move the moss API type contracts (auth, conversations) out of apps/webui
into a workspace package so desktop + webui + the moss-client SDK can all
consume one source. Repoint all 12 webui consumers to @sudowork/contracts/*.
Verified: webui client (bundler) and server (nodenext) both typecheck clean.
