+++ apps/webui/package.json
@@ -25,6 +25,7 @@
     "migrate:test": "vitest run tests/integration/db.test.ts tests/integration/session.test.ts"
   },
   "dependencies": {
+    "@sudowork/contracts": "workspace:*",
     "@arco-design/web-react": "2.66.1",
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
