+++ packages/contracts/src/auth.ts
